Drop the grid in favor of rows ala Budgie's icon-tasklist
This should make life vastly simpler because now we no longer have to deal with resizing items based on how many there are, as well as not having to worry about selection logic.

This does have the slight drawback of not being able to simply select a bunch of items and only delete or restore those, but it does still retain more fine-grained control over the items.